requirements.txt
Pillow
ultralytics
torch
